requirements.txt
apt-xapian-index==0.47
apturl==0.5.2
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
cffi==1.11.0
channels==1.1.8
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
constantly==15.1.0
cryptography==2.0.3
daphne==1.3.0
defer==1.0.6
Django==1.11.5
djangorestframework==3.6.4
feedparser==5.1.3
guacamole==0.9.2
h2==3.0.1
hpack==3.0.0
html5lib==0.999
httplib2==0.9.1
hyperframe==5.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
msgpack-python==0.4.8
oauthlib==1.0.3
onboard==1.2.0
oneconf==0.3.9
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
plainbox==0.25
priority==1.3.0
psycopg2==2.7.3.1
ptyprocess==0.5
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyOpenSSL==17.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
pytz==2017.2
pyxdg==0.25
redis==2.10.6
reportlab==3.3.0
requests==2.9.1
screen-resolution-extra==0.0.0
service-identity==17.0.0
sessioninstaller==0.0.0
six==1.11.0
software-center-aptd-plugins==0.0.0
system-service==0.3
Twisted==17.9.0
txaio==2.8.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
websocket-client==0.44.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zope.interface==4.4.3
